requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic-settings==2.1.0
beanie==1.28.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-telegram-bot==20.7
python-dotenv==1.0.0
loguru==0.7.2
aiofiles==23.2.1
tzdata==2024.1
charset-normalizer==3.3.2
